requirements.txt
python-dotenv>=1.0
django>=5.0,<6.0
dj-database-url>=2.2
psycopg2-binary>=2.9
requests>=2.31
django-cors-headers>=4.4
whitenoise>=6.7
cryptography>=43.0
requests>=2.31.0
pgvector>=0.2.5